Add smarter tile rendering, and some miscellaneous changes
-----

Added class names in the HTML for a bunch of tile parts to facilitate manual
HTML navigation.

-----

Add more view classes for each part of a tile to greatly reduce the size of and
logic in `View::Tile`.

-----

Rename tabs for tiles/tokens to "All Tiles" and "All Tokens".

-----

Logic for rendering curvilinear vs lawson track is cleaner; heavier use of
`Engine::Tile#lawson?`.

Yellow towns are temporarily rendered as lawson track so that routes to towns
can have only track that is used be highlighted. Current curvilinear track
segments are edge-to-edge; to render towns on routes properly, we'll need to
make curvilinear segments that only go half of an edge-to-edge path.

-----

Tile parts that inherit from `View::Part::Base` can set locations where they
prefer to be rendered in the function `preferred_render_locations`.

`Lib::Tile::REGIONS` defines 25 regions on a tile--the center, each edge, each
corner, and "halfway" between the center and each edge and corner.

In `preferred_render_locations`, an Array of Hashes is returned. The hashes have
`regions` and `transform` keys. `regions` defines where in the 25 regions on the
tile this part wants to be rendered; `transform` is a string that will be passed
to the SVG `transform` attribute--usually of the form `translate(x y)`.

When rendering a new part, a cost is evaluated from `@region_use` the list of
regions; whichever location in the `preferred_render_locations` Array has the
lowest cost is where the part will be rendered. Then, each of those regions has
their "cost" or "use" increased by 1 in `@region_use`.

`preferred_render_locations` can also give weighting factors to the
regions. This weighting factor affects whether the regions will be used for this
part, not how much rendering this part increases the cost in `@region_use`--that
is always 1.

In this commit, the weighting feature is only used for View::Part::LocationName;
location names want to be rendered just above center, or if that is used, just
below center, but one track down the middle is worse than two tracks at an
angle.

Consider B7/Uwajima on the 1889 map. Names can be long, so they include the
halfway points on edges 1 and 5 as regions in `preferred_render_locations`, but
we'd still prefer to render down there with the 2 track pieces than render up
top where there's only 1 track piece, so a weighting factor is applied to reduce
the cost of deciding to render below.

-----

Even if a part must always be rendered in the same spot (e.g., a segment of
lawson-style track), it should implement `preferred_render_locations` so that
its regional presence can impact where other parts get rendered.