Initial commit to track gauges using g:* notation for a path
This adds a few yellow, green, and brown tiles

In addition to the track gauges, due to the need for dashed lines,
it seemed easy to also add a bit of edging for track so there's now
a bit of white outline to each track section. This is done with a
first pass of the svg that's slightly wider than the actual track,
and a contrasting color.

This outline helps a bit to distinguish crossings, as on tile #16
and #20. This is due to rendering each outline + path as a unit.

However, joint edges on #24 and #43 have a bit of oddness
to them, which would be helped by rendering the underlying outline
all at once, then rendering the track later.