Attempt to move phase change after action
This is a partial implementation of moving the phase change after
the processing of a round action. The new conditional method calls
are needed because the current logic expects new phase rules to be
in place during the processing of an action. Unfortunately there
are six test failures where existing games run into a player
acting out of turn, due to turn order not being appropriately
advanced. Further investigation is required.

resolves #332