change upgrades_to? specs to one test per game
These specs were originally written as one test per upgrade to ensure the rspec
output would be useful, but the number of tests that created is not so useful.

Using aggregate_failures to group upgrades_to? tests by game gets the number of
tests to something more reasonable.

Using expected/actual strings instead of just the return value of
`upgrades_to?` (`true`/`false`) keeps the rspec output useful.

For these tests, `A < B` means "A upgrades to B" and `A !< B` means "A does not
upgrade to B"