Improve game end: current OR and/or stock market reach max (#1101)
* Improved par price order so that it is always ascending

The par price when paring corporations did not appear in
correct order.

* Refactor game end check

Add a GAME_END_CHECK that can be used to configure game end check
per 18xx.

Implement support for handling game end in current OR.

Renamed current_round/full_round to current_or/full_or in case
we need to support end game in SR.

Ensure after is taking precendence to reason, so that all
reasons are checked for immediate, before current OR etc.

* Introduce a game end check for stock market reaching max

Add a possibility to configure stock market prices as end game
triggers by adding "e" to them. This appear as blue boxes in the
stock market.

If end game stock market reached a flag is raised that can
be checked by the end game check.

Added reason :stock_market in end game check that will check
the stock market end game situation if this has been configured.
By default reaching stock market end game boxes will not
trigger end game.

* Introduce a game end check for a set number of OR sets.

If :final_or_set is set to a number, this will be the last
OR set in the game.

* 18AL/18GA/18MEX/18MS/18TN: Apply end game checks

18AL/18GA/18MEX/18TN has three game end situations:
- Bankrupcy, end directly
- Bank run out, current OR (or OR after current SR)
- Stock market price reach top, current OR

Stock market prices that trigger game end:
- 18MEX $200
- 18AL, 18GA and 18TN $300

For 18MS the only end game cause is when OR set 10
is completed.