Display corporation abilities on the charter
Add two new accessors to the ability base class:
 - name
 - display_value
These can be set during initialization, or later.

When displaying corporation charter a section for all
abilities a company has that is not part of any company.
The name and display_value is displayed for all these abilities.

Does not support showing more information when clicked at.

Note! Only abilities that are owned by the corporation are displayed.
Abilities owned by a company owned by the corporation are
not displayed.