18AL: Correct Montgomery tile
Added handling of tile lay in Montgomery to ensure that
the tile will have the label 'M' to restrict the green and
brown tiles to M tiles. From the start Montgomery does not
have any label to allow for any yellow city tiles to be laid
there.

This change was part of an earlier commit but seems to have
been lost during migration.