18AL: Add support for historical objective (#961)
Configured historical objective hex for all corporations in 18AL
using the assign_hexes ability. When a corporation places a
token in its historical objective hex, it receives $100 to treasury
as it has reached its historical objective.

For historical objectives attributes the name is "Historical
objective" and the display_value is set to the city name of
the destination. This is shown in the corporation charter.

When historical objective reached the ability is removed, which
means that it disappear from the corporation charter.
A descriptive log is produced so that the player will notice
why the company got $100.