18AL: Add support for historical objective (#961) (#1050)
* 18AL: Correct Montgomery tile

Added handling of tile lay in Montgomery to ensure that
the tile will have the label 'M' to restrict the green and
brown tiles to M tiles. From the start Montgomery does not
have any label to allow for any yellow city tiles to be laid
there.

This change was part of an earlier commit but seems to have
been lost during migration.

* Refactor 4D revenue modification as a general method

Extract into a method that can be included from other 18xx.

* Refactor terminus check to be an includable method

* 18MEX: Apply 4D revenue and terminus check (Merida)

* Display corporation abilities on the charter

Add two new accessors to the ability base class:
 - name
 - display_value
These can be set during initialization, or later.

When displaying corporation charter a section for all
abilities a company has that is not part of any company.
The name and display_value is displayed for all these abilities.

Does not support showing more information when clicked at.

Note! Only abilities that are owned by the corporation are displayed.
Abilities owned by a company owned by the corporation are
not displayed.

* 18AL: Add support for historical objective (#961)

Configured historical objective hex for all corporations in 18AL
using the assign_hexes ability. When a corporation places a
token in its historical objective hex, it receives $100 to treasury
as it has reached its historical objective.

For historical objectives attributes the name is "Historical
objective" and the display_value is set to the city name of
the destination. This is shown in the corporation charter.

When historical objective reached the ability is removed, which
means that it disappear from the corporation charter.
A descriptive log is produced so that the player will notice
why the company got $100.