Add a discount field in the tile_lay ability
It is now possible to configure a tile_lay ability to apply
a discount to the cost for laying the tile. By default this
is 0 (so no change).

The discount is needed as some extra tile lays still incurs
some cost instead of being free or full, e.g. 18MEX.