18AL: Implement company Brown & Sons Lumber Co (#1073)
* Override special tile laying except color

If tile laying allows you to lay a tile it should be possible
even if terrain does not match.

This commit lets you override everthing but color restriction when
laying tiles via an ability.

* Add a discount field in the tile_lay ability

It is now possible to configure a tile_lay ability to apply
a discount to the cost for laying the tile. By default this
is 0 (so no change).

The discount is needed as some extra tile lays still incurs
some cost instead of being free or full, e.g. 18MEX.

* 18AL: Implement company Brown & Sons Lumber Co

Add ability for Brown & Sons Lumber Co to lay tile 445  as a
one of ability, for free.

Tile 445 does now ignore town count in the hex it is placed
via the Brown & Sons Lumber Co special ability.

* 18MEX: Implement company Kansas City, Mexico, & Orient Railroad

Add ability for Kansas City, Mexico, & Orient Railroad to lay tile
470 via its special ability. There is a discount of $60 to lay this
tile so the cost will be $60 instead of the full cost of $120.