18MEX: Correct configuration and add phase related events.
Correct quite a lot of configuration for trains and phases.

Split 3 train (6) to 3 train (4) and 3½ train (2).

Split 6 train (2) to 6 train (1) and 6½ train (1).

Change 4 train to rust on 6½ train purchase.

Add phase 3½, which is a copy of phase 3, but with a close
minors event. The event is still a dummy event.

Add a dummy NdM merger event for phase 5.

Add phase 6½, which is a copy of phase 6. Used to rust 4 trains.

Add event description for events specific to 18MEX.
Add trivial event handling methods - they just write out
something and have no other effect as of now.

Rename CHI logo file to match the naming of the other logo files.