Fix/issue 1157 (#1158)
* Improve presentation of corporation abilities

Remove display_value field from attribute. Use name instead but
rename it to description to better fit its usage.

Do not show corporation ability on corporation charter if it has
no description.

Make it possible to show more information, and adjust dynamically.

* 1846: Fix corporation abilities presentation (#1157)

Assign description to the corporation abilities that should
be shown in the corporation charter. Some abilities is more
for internal reasons and does not need to be shown.

* 18AL: Update description for historical objective

Need to update 18AL so it conforms to the way abilites are shown.