18AL: Implement company South & North Alabama Railroad
The S&NA/SNAR company has a one time ability to select
one of 5 hexes to put out a token that gives a $10 bonus
to any of its routes that uses that hex. The token remains
even after company is closed, but the token is removed
when phase 6 is entered.

This is implemented as a assign_hex ability in SNAR, which
when triggered adds a hexes_bonus ability to the owning
corporation of $10 for the assigned hex. The ability is
removed when phase 6 is entered.

Also note that the corporation need to own a train, and
have access to the city in question to be able to assign
the token there.

The assigned hex is shown as a corporation ability.