18AL: Implement company Memphis & Charleston Railroad
The M&C RR company has an ability that if one of the routes
contain a specific pair of cities then a bonus is received.
The bonus is received once per OR.
Their are two pairs of such bonus cities, and they have
different bonuses:
 - Atlanta-Birmingham, $20 (Robert E Lee)
 - Mobile-Nashville, $40 (Pan American)
This ability persist for the whole game, except in the
case of change of presidency - then the ability is removed.

The implementation of this ability is done in the following
way:
 - When the M&C RR company is bought, two hexes_bonus abilities
   are added to the purchasing corporation
 - The route bonuses ability is represented by two instances
   of the hexes_bonus ability which has the pair of cities
   for that specific bonus, the requirement that all (both)
   hexes must be passed in a route to be applicable and the
   requirement that this bonus can only be applied once per OR.
 - If the corporation with this abilities changes president
   both bonuses are removed. As the ability is in the
   corporation, and not in a company, closure of company does
   not affect the abilities.
 - The route bonus abilities are shown as two lines in the
   corporation charter.