Introduce special train buy based on train_discount ability
Added a train_discount ability that makes it possible to make
discounted train buys, either by absolute amount or buy a
fracitional amount.
The discounted price is shown for applicable trains if
ability is activated.

Have split Step::Train into a Step::BuyTrain class and a
Step::Train module. The module is included by the class.
This is to simplify subclassing for train buys.

Have also renamed the Train step to BuyTrain, which meant
chaning quite a few files.