18AL: Implement M&C RR and S&NA companies (#1089)
* Introduce a hex bonus ability

Used to handle abilities that gives bonuses based on which
hexes are included in the route(s).

* Make it possible to modify revenue based on all routes

Introduce a routes_revenue method in Engine::Game::Base that
by default just return the sum of the revenue for each route.

This makes it possible to make modifications based on the
complete route picture, which means that the dividend used
in the view can be modified.

* 18AL: Refactor - extract method to get location name of hex

* 18AL: Implement company South & North Alabama Railroad

The S&NA/SNAR company has a one time ability to select
one of 5 hexes to put out a token that gives a $10 bonus
to any of its routes that uses that hex. The token remains
even after company is closed, but the token is removed
when phase 6 is entered.

This is implemented as a assign_hex ability in SNAR, which
when triggered adds a hexes_bonus ability to the owning
corporation of $10 for the assigned hex. The ability is
removed when phase 6 is entered.

Also note that the corporation need to own a train, and
have access to the city in question to be able to assign
the token there.

The assigned hex is shown as a corporation ability.

* 18AL: Implement company Memphis & Charleston Railroad

The M&C RR company has an ability that if one of the routes
contain a specific pair of cities then a bonus is received.
The bonus is received once per OR.
Their are two pairs of such bonus cities, and they have
different bonuses:
 - Atlanta-Birmingham, $20 (Robert E Lee)
 - Mobile-Nashville, $40 (Pan American)
This ability persist for the whole game, except in the
case of change of presidency - then the ability is removed.

The implementation of this ability is done in the following
way:
 - When the M&C RR company is bought, two hexes_bonus abilities
   are added to the purchasing corporation
 - The route bonuses ability is represented by two instances
   of the hexes_bonus ability which has the pair of cities
   for that specific bonus, the requirement that all (both)
   hexes must be passed in a route to be applicable and the
   requirement that this bonus can only be applied once per OR.
 - If the corporation with this abilities changes president
   both bonuses are removed. As the ability is in the
   corporation, and not in a company, closure of company does
   not affect the abilities.
 - The route bonus abilities are shown as two lines in the
   corporation charter.