18AL: Fix problems with revenue calculation
Contained an incorrect attribute key.

Also has changed to simplify the calculation a bit. Route bonus
now only have to be calculated when showing to total revenue, and
are not included in the revenue per route. This means that the
total revenue will be the sum of revenues per route plus a one
time addition per each route bonus any route has triggered.

Finally, corrected the Withhold/Payout total shown, so that
the @game.revenue_for method is used to possibly modify the
total revenue for the routes.