18AL/18MS: Implement company that gives discount for some trains (#1074)
* 18AL: Fix problems with revenue calculation

Contained an incorrect attribute key.

Also has changed to simplify the calculation a bit. Route bonus
now only have to be calculated when showing to total revenue, and
are not included in the revenue per route. This means that the
total revenue will be the sum of revenues per route plus a one
time addition per each route bonus any route has triggered.

Finally, corrected the Withhold/Payout total shown, so that
the @game.revenue_for method is used to possibly modify the
total revenue for the routes.

* Refactor to allow introduction of special train buy

By storing bought trains in round state, different instances
of train buy can ensure correct upper limit of train buys
from depot are enforced.

* Introduce special train buy based on train_discount ability

Added a train_discount ability that makes it possible to make
discounted train buys, either by absolute amount or buy a
fracitional amount.
The discounted price is shown for applicable trains if
ability is activated.

Have split Step::Train into a Step::BuyTrain class and a
Step::Train module. The module is included by the class.
This is to simplify subclassing for train buys.

Have also renamed the Train step to BuyTrain, which meant
chaning quite a few files.

* 18AL: Implement New Decatur Yards company

New Decatur Yards ability is to give a choice of using
a discount for purchasing one train.

The discount is set to 50% for 3T, 4T and 5T.

When train bought for a discounted price the company closes
(if count is zero)

* 18MS: Implement Mobile & Ohio company

Implement train discount for Mobile & Ohio Railway,
100$ discount for 3+, 4+ trains