Show status connected to each phase (#1325)
Use a similar method as events, to show status that can be
active during a phase. E.g. Can Buy Trains are now shown
with a description in a status list after the phase.

Intention is to be able to add more status per game, e.g.
train buy limit during phase 2 and phase 3 for 18AL/18MEX.