18AL: Add support for buying companies between players
During an SR a player can buy player owned companies
from other players. If the active player still has
buy left for the stock turn, an entry field and one
or more buttons (for all selected player owned companies
that other players own). When player click on one of the
buttons the company is bought from the other player for
the specified amount.

The purchase is automatic - no confirmation. A bold log
is done to signal to the players that they might want
to verify, and possibly undo, if it was not an agreed
upon sale.

This feature is only active if event status
can_buy_company_from_other_player is set for the current
phase. Also, the buy_sell_par_shares.rb need to be
subclassed/modified for the game, to support the company buy.