18TN: Correct OR and phases
Make some corrections to fix a bug that made
it got stuck after SR 1.

Split 3 trains into a group of 3 and a group of 3'.
Add Phase 3½ which is triggered by purchase of a 3' train.

Split 6 trains into one 6 and one 6', and make
Add Phase 6½ which is triggered by purchase of the 6' train.