18TN: Add Civil War event
Add a Civil War event which is triggered by phase 3½.

A company that has train(s) when Civil War is triggered will
receive a civil_war ability. During its next OR one of the
trains run will loose its revenue. (Implemented as the train
that gives lowest revenue [including 0 as in not run] will
give effective revenue 0.

After applying the civil war on a corporation's OR run it
is removed.