18TN: Special handling needed for Chattanooga, Memphis, Nashville
Chattanooga now starts on the board as a yellow 0 revenue C
hex. When upgraded to green C tile, the tile changes to P tile
to allow only for brown P upgrades.

Nashville is similar to Chattanooga, but starts with a N tile,
upgrades to green N, and is changed to P to allow for brown
P upgrade.

Memphis is a normal tile, but when upgrading to brown it now
allows either normal brown or a P labeled tile.

A method label_check_override which can be implemented in a
game that need to handle the label check by its own.
In case of 18TN this is used for the Memphis upgrade.