Add two bold texts, and improve phase status (#1329)
* 18AL: Make important message bold

The "Note!" message connected to the assign of Warrior Coal
Field token is now made bold, to increase the chance for the
players to note it.

* Make "Game over" message in log bold

* Show status connected to each phase (#1325)

Use a similar method as events, to show status that can be
active during a phase. E.g. Can Buy Trains are now shown
with a description in a status list after the phase.

Intention is to be able to add more status per game, e.g.
train buy limit during phase 2 and phase 3 for 18AL/18MEX.

* Use the phase status mechanism for 18AL/18GA/18MEX/18TN

Add status text for phase 2 and phase 3 about limited
train buy.