18MS: Fix map to make it playable with all hexes available. (#1385)
Corrected off board revenue for New Orleans.

Make a three hex off board group, with correct revenue, in Atlanta.

Make a real town hex that is connected to New Orleans. Put this
in J2 to work around that you cannot have town and city in the
same hex. Name it Port Sulphur after a real port south east of
New Orleans. (Or should it use another name? Did not see any
on the 18MS map.)