Make 18TN playable (#1383)
* 18TN: Allow 2 yellow or 1 upgrade every OR

* 18TN: Correct OR and phases

Make some corrections to fix a bug that made
it got stuck after SR 1.

Split 3 trains into a group of 3 and a group of 3'.
Add Phase 3½ which is triggered by purchase of a 3' train.

Split 6 trains into one 6 and one 6', and make
Add Phase 6½ which is triggered by purchase of the 6' train.

* 18TN: Add Civil War event

Add a Civil War event which is triggered by phase 3½.

A company that has train(s) when Civil War is triggered will
receive a civil_war ability. During its next OR one of the
trains run will loose its revenue. (Implemented as the train
that gives lowest revenue [including 0 as in not run] will
give effective revenue 0.

After applying the civil war on a corporation's OR run it
is removed.

* 18TN: Implement private company LNR

Closes when pared (which happens when LNR bought)

* 18TN: Correct 4 tile laying privates

* 18TN: Special handling needed for Chattanooga, Memphis, Nashville

Chattanooga now starts on the board as a yellow 0 revenue C
hex. When upgraded to green C tile, the tile changes to P tile
to allow only for brown P upgrades.

Nashville is similar to Chattanooga, but starts with a N tile,
upgrades to green N, and is changed to P to allow for brown
P upgrade.

Memphis is a normal tile, but when upgrading to brown it now
allows either normal brown or a P labeled tile.

A method label_check_override which can be implemented in a
game that need to handle the label check by its own.
In case of 18TN this is used for the Memphis upgrade.

* 18TN: Make Illinois Central presidency share 30%

* 18TN: Add wiki link on the game info page

* 18TN: Support corporation purchase of company during OR1

Corporations can buy companies for face value during OR1.

After that purchase can only be done in phase 3 and phase 4
and then for 50%-150% of face value.

* Fix a review comment

* Corrections of bugs found in 18TN implementation

* 18TN: Add first spec

* Refactor upgrades_to? to be in game instead of tile

Allows for a title to have control of upgrades.

18TN now uses this to control upgrade to brown in Chattanooga,
Memphis and Nashville.