Handle empty depot smothly (#1446) (#1451)
If the last train is bought the game does not try to use
the cheapest train in the depot.

And if no trains are in the depot then a corporation is not
forced to buy trains.

Included a test case for an 18AL game where all 9 permanent
trains are bought, and the trains distributed among the 6
corporations so that one corporation is left without. Game
runs to completion with L&N trainless.