18TN: Handle special case in Civil War event (fixes #1457)
If a corporation has the Civil War effect applied, and it only
has one train, the company cannot pay out. So this situation
will not move the stock price. This is an exception to the
normal rules for Payout/Withhold.

Have changed the view so that it shows a button "Civil War"
and the stock change will be "None". Also improved the logging
for this case.