18TN: Bug fixes (#1445)
* 18TN: Tile corrections (#1427, #1441)

Make Nashville a white hex, without any label, from the start.
When upgrading Nashville to yellow put an N label on the hex
to make green upgrade to N tile obvious.

In the tile manifest, show brown P tile as upgrade possibility
for green N, green C and green cities.

When showing upgrade tiles for a hex, pass hex to ensure that
any location based tiles do not appear for other locations.

* 18TN: Allow corporation to buy at most one company in OR 1 (#1442)

* 18TN: Handle special case in Civil War event (fixes #1457)

If a corporation has the Civil War effect applied, and it only
has one train, the company cannot pay out. So this situation
will not move the stock price. This is an exception to the
normal rules for Payout/Withhold.

Have changed the view so that it shows a button "Civil War"
and the stock change will be "None". Also improved the logging
for this case.

* 18TN: Update spec test

* Fix review comment about all_potential_upgrades

Tile knows its hex if it is map context so no need to pass hex
explicitly