18AL: Fix how Montgomery green tile is handled (Fixes #1473)
Previously we added an M label to the yellow tile placed in
the Montgomery hex, but this had negative effects when tile is
removed, as label was not cleaned up. (Which made yellow tile
fail for later upgrades in other hexes.)

Instead the 18AL game take control of upgrade to green tile in
Montgomery, and also (as before) give green Montgomery tile as
one of the potential upgrades of yellow city tiles in the tile
manifest.