18TN: Improve tile handling (fixes #1476, #1477)
Memphis can now only be upgraded to brown P tile.

Do not add any labels when upgrading tiles, as this will
cause problems if not cleaned up correctly.

Have refactored the upgrades_to? and all_potential_upgrades
to handle tile manifest better.

About #1477 the gray tile is shown when clicking on brown P
tile on the board, but it is grayed out and have the text
"Later Phase" which is the normal result.

Have also refactored 18AL/18GA to use the same style of
tile handling. And removed label adding code to 18GA to
avoid the bug this causes.

Have added an optional parameter to all_potential_upgrades.
As super class does not use this it has been disabled in
rubocop. There is a Lint/UnusedMethodArgument:
  AllowUnusedKeywordArguments: false
that could be turned on instead, but disable all UnusedMethodArgument
is fine on that line.