18MS: Correct handling of free train (Fixes #1536) (#1540)
The free train received when buying private Ocilla Souther RR
did not have the correct index. Changed so that the free train
is created as all other 2 trains, and then bought by neutral
corporation and set to buyable false. When Ocilla Southern RR
is bought the owning corporation makes a free buy of the free
2 train. The 2 train is set to not buyable otherwise so that
the owning corporation cannot sell it later.