18MS: Add New Orleans bonus
First to token in New Orleans gets a one time corporation bonus
of $100. And also receives a route bonus of $10 for any route
include New Orleans. The route bonus are removed when the first
6 train is bought.

A coins icon has been added on the New Orleans hex. When bonus is
triggered the icon is removed.