Make it possible to use special tile_lay without specifying hex/tiles
If hexes is not specified, special tile_lay is possible in any
hex (which connects to existing tracks for the corporation).

If tiles is not specified, special tile_lay is possible for any
tile that is allowed as normal tile lay in that hex.

If tile_lay ability sets reachable attribute to true, a check will
be done that the new tile creates a reachable track to one of
controlling corporation's station.