18MS: Implement tile laying privates P1 and P2
P1 can make special lay of one yellow tile. Any terrain costs are
ignored.

P2 is similar to P1 but can lay two special tile lays, both for free.

The special tile lays can be done before, after or in between of
normal tile lay or upgrade.

The tile lay ability is removed after tile lay completed, even if
only one tile was laid by P2.

The player must click on the private to activate the special tile
lay. For P2 this need to be done before each of the lays.

One problem exists, but this might be OK. When ability is selected
no hexes are high lighted. It is possible to try to lay a tile
anywhere on the board, but the UI reports a game error (not
connected) if lay does not make the hex connected to the acting
corporation.