18MS: Assisted permanent train buy
If a corporation owns trains but cannot afford to buy an available
permanent train (but has at least 50% of its price) the president
can assist in the purchase buy spending the difference (and an
extra $50 "fee").

Have added a president_assisted_buy in the base class that can
be overridden if the game support this feature.