Phase change should occur when buy of train is complete (fixes #1553)
If phase change is done to early this might result in a cheaper train
for a later phase becomes available which caused the buy_train_action
to bug out.

Instead, do the phase change when all preparations and checks has been
done, just before the actual buy. But do put out the logs of the buy
after a potential phase change.