18MS: Handle emergency buy with insufficient resources
In 18MS a player that is president for a corporation doing an
emergency buy cannot declare bankrupcy. Instead the player
must buy the cheapest train in the depot and borrow the
amount missing (+ $50 interest).

If player has negative cash at the end of an SR a 50% interest
of the remaining debt are added to the debt.

If the game does not support and step that can handle bankruptcy
the "Declare Bankruptcy" is not shown during emergency buy.
In 18MS you cannot declare bankruptcy - you need to manually
end the game if a player has given up.