Enforce that extra tile lays is actually part of consecutive tile lay
Fixes #1583

An extra tile lay (that is, tile_lay ability, with when='track)
must be preceeded and/or followed by a tile lay action
or be done during a Track or TrackAndToken step.

Only exception is that Pass or BuyCompany is ignored for this check.
Undo and Redo is also handled so that canceled actions are ignored
for the check.

Corrected two spec test games where the extra tile lay was done
outside of the consecutive tile laying. The correction was to
move the extra tile lay action so it became legal.