18MS: Implement Meridian & Memphis Railway private (#1548)
* 18MS: Implement Meridian & Memphis Railway private

This private has a one time ability (for a corporation) to place
a token at half price. A similar way of specifying this as for
train_discount is used, and a discount attribute has been added
to the token ability. The discount is how much (0 to 1) should
be deducted from the regular price for a token placed, so e.g.
0.75 would mean a 75% discount.

Use similar method as was done in 18GA Waycross & Southern to
just use ability if ability is actived explicitly.

Have made a change to hex selection. If hexes is empty this
means all (connected) hexes is allowed.

Use of ability does not close the private.

* After review comments

Also avoid showing special token ability if outside standard token
step.