18MS: Handle emergency buy (#1559)
* Phase change should occur when buy of train is complete (fixes #1553)

If phase change is done to early this might result in a cheaper train
for a later phase becomes available which caused the buy_train_action
to bug out.

Instead, do the phase change when all preparations and checks has been
done, just before the actual buy. But do put out the logs of the buy
after a potential phase change.

* 18MS: Handle emergency buy with insufficient resources

In 18MS a player that is president for a corporation doing an
emergency buy cannot declare bankrupcy. Instead the player
must buy the cheapest train in the depot and borrow the
amount missing (+ $50 interest).

If player has negative cash at the end of an SR a 50% interest
of the remaining debt are added to the debt.

If the game does not support and step that can handle bankruptcy
the "Declare Bankruptcy" is not shown during emergency buy.
In 18MS you cannot declare bankruptcy - you need to manually
end the game if a player has given up.

* Fix review comment

* Modify emergency_money after review comment