18MS: Handle emergency buy with insufficient resources
In 18MS a player that is president for a corporation doing an
emergency buy cannot declare bankrupcy. Instead the player
must buy the cheapest train in the depot and borrow the amount
missing (+ $50 interest).

Debt is represented by negative cash (which is an existing
"feature). This stops the debtee to spend any cash (except
further loans)

If player has negative cash at the end of another loan is added
to the debt, which means that the debt is increased by 50% of
the remaining debt.

The final score in the game is share value plus cash, so this
means that any remaining debt after OR5.2 is automatically
deducted from the final score.

The display of the "Declare Bankruptcy" button during emergency
buy has been changed. If the active step(s) does not support
the 'bankrupt' action the button is not displayed.
In 18MS you cannot declare bankruptcy - you need to manually
end the game if a player has given up, so therefor no button.

Have removed automatic phase change in 18MS for 6 and D trains.
Instead this is checked directly after buy train action in
18MS module. The reason for this is to avoid the problem that
a cheaper train made available by the phase triggered by trying
to buy the train of the new phase. This is reported in issue
Now the phase change in 18MS does not rust any trains so doing
phase change when buy train completed should be an OK way of doing
it in 18MS.
Note that doing phase change to D (gray) would probably not
cause any problems but it is done in the same place in the code
as change to brown (6) phase to make it consistent.
From before the phase change to green (3) is handled in g_18_ms.rb
when we enter OR1.2.

Also added a check that the cheapest variant of a train is bought
during emergency buy; in case of 18MS it was possible before to
emergency buy the 4D instead of 2D.