18MS: Implement private P1 and P2 + general tile lay anywhere (#1550)
* Make it possible to use special tile_lay without specifying hex/tiles

If hexes is not specified, special tile_lay is possible in any
hex (which connects to existing tracks for the corporation).

If tiles is not specified, special tile_lay is possible for any
tile that is allowed as normal tile lay in that hex.

If tile_lay ability sets reachable attribute to true, a check will
be done that the new tile creates a reachable track to one of
controlling corporation's station.

* 18MS: Implement tile laying privates P1 and P2

P1 can make special lay of one yellow tile. Any terrain costs are
ignored.

P2 is similar to P1 but can lay two special tile lays, both for free.

The special tile lays can be done before, after or in between of
normal tile lay or upgrade.

The tile lay ability is removed after tile lay completed, even if
only one tile was laid by P2.

The player must click on the private to activate the special tile
lay. For P2 this need to be done before each of the lays.

One problem exists, but this might be OK. When ability is selected
no hexes are high lighted. It is possible to try to lay a tile
anywhere on the board, but the UI reports a game error (not
connected) if lay does not make the hex connected to the acting
corporation.

* Document new reachable attribute, and add missing documentation for hex_bonus

* Correction after review comment