Cannot contribute money if affordable trains exists
Correct a bug; in case a player tries to buy a train that the
corporation cannot afford, but at the same time there is a
train it can afford, an error message now appear explaining
this fact. This changes does allow for the player to do
buy the non-cheapest train, if it cannot afford the cheapest.