Fix/18 ms/abilities (#1713)
* 18MS: Correct stock round

Stock round in MS only allows buy and sell shares. No ability can
be used during stock round.

Having special track in stock round caused tile laying abilities to
appear in stock round.

* 18MS: Fix Chattanooga bonus (fixes #1698)

The check was done in lay_tile method but that was not available for
when Chattanooga reached by special tile. So have refactored so
that this check is available for both normal and special lay.
This also resulted in a change in the spec test game as the player
now had to spend $50 less to buy a train.