18MEX: Support NdM merge
When first 5 train is bought the NdM merge event is triggered
if NdM has floated.

The merge process is that player in order, starting with the
player to the left of the player that did buy the triggering
5 train, is asked if their corporation wants to merge, one
candidate corporation at a time. (Excluding "American" corporations,
and corporations controlled by the NdM president.)

If there is no floated corporation that accepts, NdM will try
to merge with any non floated, if there is any eligible one.
Non floated corporations that do not have a player owned
presidency share is only choice if such exists.
If there are several non floated candidates, NdM president
gets to choose.
If there is just one non floated candidate, this is selected
automatically.

The owner of president share of a merged corporation receives
a 10% reserved NdM share in exchange for it. Remaining shares
in merged corporation is force sold for half market price.

Treasury and trains are transfered to NdM.

If merge occur, NdM receives to exchange token (cost $80).
NdM also exchanges up to two exchange tokens, one should
at least be the home token if any. If there are more than
three tokens out on the board (from the corporation to
merge) NdM gets to choose which to exchange. Any other
tokens are removed from the board.
Note! If a candidate token is located in a city there NdM
already has a token, this candidate token is removed.

If no corporation accept, or if NdM has not floated, there will
be no merge, and NdM will not get any exchange tokens, and the
certificate limit is increased by one. The trade-in share
is instead made into a normal buyable share.
Note! This does not affect +1 train limit, or train buy/sell
are not affected by the failure to do any merge.

There are a new step: merge, which is located in module 18MEX.
This is as it works better to have it as a step in the
ongoing operation round, as the operation round will continue
when the merge has completed.
This is used for the merge question. The merger view used by
1817 today has been enhanced a bit to work with 18MEX.

18MEX also uses a new assign step which is used in case NdM need
to choose between several tokens on the map.

---------

Present the reserved (non buyable) shares in the corporation charter.
For 18MEX it is shown as Trade-in shares.